Powerups (1/2) Created Powerups.cs: Finds GameObject with tag "Player" and we use it to GetComponent on health, PlayerShooting, & movement. If the player triggers a collider on a powerup that equals the tag we handle it. Afterwards it destroys the object. Powerup abilities added: Heal to full, Invincibility, Speed, DamageBoost.
Health is just set to starting health
Invincibility is harvested from the feature that if a player is hit, they will be invincibile for 1.3 seconds to adjust themselves.
Speed is mulitiplied from a random number 1.0 - 3.0
DamageBoost is double the damage and 2-3X fire rate.

At the end of the powerup session the values need to be reset back to "normal". For this reason several scripts were edited to make this happen. DamageBoost has to know the current weapon to know what the "set of damage/hits per second" is.

Created Powerupmanager.cs: Spawns powerups in random spots. Spawn time of powerups is currently only randomized once per game between 5-20 and I'd like to change this later. Powerup type is randomized and not all types of powerups will be available at once, it all just depends on the randomizer.