Minor, but major change on PlayerShooting:
After writing up the last commit I want to do the damage powerup slightly different. It seems like a sure bet and the most OP powerup. So to mix things up a bit the damageMultiplier & speedMultiplier will be a random number from .5 to 3, meaning it could actually hinder your current state for the duration.